fix: move the tests of the moderation module
